include/boot/s390.h: Always use __vector128 from asm/types.h
asm/types.h may be included after the guard. This happens eg. when
compiling with musl libc. By including the header directly we do not
rely on the order of includes and can avoid the need to define
__vector128 ourselfs.

Alternatively it would be possible to use a different name than the
kernel headers to avoid the need for the guard.

Ref 11bdab26297e include/boot/s390.h: add guard for `struct __vector128`

Signed-off-by: Sertonix <sertonix@posteo.net>
Reviewed-by: Steffen Eiden <seiden@linux.ibm.com>